pyproject.toml
[tool.poetry]
name = "readverse"
version = "0.1.0"
description = ""
authors = ["Ren <renrror@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-admin = "^1.6.1"
flask-migrate = "^4.0.5"
flask-login = "^0.6.3"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
